% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_active-voice.R
\name{addin_active_voice}
\alias{addin_active_voice}
\title{active voice Addin}
\usage{
addin_active_voice()
}
\description{
Call this function as a Rstudio addin to ask GPT to change selected text
into the active voice
}
