% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_chat_gpt.R
\name{addin_chatgpt}
\alias{addin_chatgpt}
\title{Run Chat GPT
Run the Chat GPT Shiny App}
\usage{
addin_chatgpt()
}
\value{
This function has no return value.
}
\description{
Run Chat GPT
Run the Chat GPT Shiny App
}
\examples{
# Call the function as an RStudio addin
\dontrun{addin_chatgpt()}
}
