% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_comment-code.R
\name{addin_comment_code}
\alias{addin_comment_code}
\title{active voice Addin}
\usage{
addin_comment_code()
}
\description{
Call this function as a Rstudio addin to ask GPT to add comments to your code
}
