% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_spelling-and-grammar.R
\name{addin_spelling_grammar}
\alias{addin_spelling_grammar}
\title{spelling and grammar Addin}
\usage{
addin_spelling_grammar()
}
\description{
Call this function as a Rstudio addin to ask GPT to improve spelling and
grammar of selected text.
}
