% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{ask_to_set_api}
\alias{ask_to_set_api}
\title{Ask to set API key}
\usage{
ask_to_set_api(try_again = FALSE)
}
\arguments{
\item{try_again}{A boolean indicating whether the user has already tried to
set the API key.}
}
\description{
This function asks the user if they would like to set the API key.
}
