% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_taglist.R
\name{attrs_to_params}
\alias{attrs_to_params}
\title{Convert Attributes to Parameters}
\usage{
attrs_to_params(attrs)
}
\arguments{
\item{attrs}{A named character vector representing attributes}
}
\value{
A character string that can be used as a parameter list in an HTML
tag
}
\description{
This function takes a named character vector representing attributes and
returns a character string that can be used as a parameter list in an HTML
tag.
}
