% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-transcribe-audio.R
\name{audio_to_data_uri}
\alias{audio_to_data_uri}
\title{Convert Audio File to Data URI}
\usage{
audio_to_data_uri(file_path)
}
\arguments{
\item{file_path}{A string. The path to the audio file.}
}
\value{
A string containing the data URI.
}
\description{
This function takes an audio file path and converts it to a data URI.
}
