% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_dropdown.R
\name{bs_dropdown}
\alias{bs_dropdown}
\title{Boostrap 5 dropdown}
\usage{
bs_dropdown(label, id, ..., class = NULL)
}
\arguments{
\item{label}{The label}

\item{id}{ID of the dropdown element}

\item{...}{additional arguments passed to card_body()}

\item{class}{Additional classes}
}
\value{
A Bootstrap 5 dropdown
}
\description{
Boostrap 5 dropdown
}
