% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_perform_request.R
\name{Buffer}
\alias{Buffer}
\title{Message Buffer}
\description{
A message buffer
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Placeholder for the message}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Buffer-add_chunk}{\code{Buffer$add_chunk()}}
\item \href{#method-Buffer-clone}{\code{Buffer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Buffer-add_chunk"></a>}}
\if{latex}{\out{\hypertarget{method-Buffer-add_chunk}{}}}
\subsection{Method \code{add_chunk()}}{
Add a message chunk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Buffer$add_chunk(chunk)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chunk}}{A message chunk}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Buffer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Buffer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Buffer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
