% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prompt.R
\name{chat_create_history}
\alias{chat_create_history}
\title{Chat history}
\usage{
chat_create_history(response)
}
\arguments{
\item{response}{A response from \code{gpt_chat()}.}
}
\value{
list of chat messages
}
\description{
This takes a response from chatgpt and converts it to a nice and consistent list.
}
