% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_queries.R
\name{chat_create_system_prompt}
\alias{chat_create_system_prompt}
\title{Create system prompt}
\usage{
chat_create_system_prompt(style, skill, in_source)
}
\arguments{
\item{style}{A character string indicating the preferred coding style, the
default is "tidyverse".}

\item{skill}{The self-described skill level of the programmer,
default is "beginner"}

\item{in_source}{Whether to add intructions to act as in a source script.}
}
\value{
A string
}
\description{
This creates a system prompt based on the user defined parameters.
}
