% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prompt.R
\name{chat_history_append}
\alias{chat_history_append}
\title{Append to chat history}
\usage{
chat_history_append(history, role, content)
}
\arguments{
\item{response}{A response from \code{gpt_chat()}.}
}
\value{
list of chat messages
}
\description{
This appends a new response to the chat history
}
