% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{check_api}
\alias{check_api}
\title{Check API setup}
\usage{
check_api()
}
\value{
Nothing is returned. If the API key is valid, a success message is
printed. If the API key is invalid, an error message is printed and the
function aborts.
}
\description{
This function checks whether the API key provided in the \code{OPENAI_API_KEY}
environment variable is valid. This function will not re-check an API if it
has already been validated in the current session.
}
\examples{
# Call the function to check the API key
\dontrun{
check_api()
}
}
