% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{check_api_connection}
\alias{check_api_connection}
\title{Check API setup}
\usage{
check_api_connection(api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{api_key}{An API key. Defaults to the value of the `OPENAI_API_KEY` environment variable.}
}
\value{
Nothing is returned. If the API key is valid, a success message is printed. If the API key is invalid, an error message is printed and the function aborts.
}
\description{
This function checks whether the API key provided in the `OPENAI_API_KEY` environment variable is valid.
}
\examples{
check_api_connection()
}
