% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptstudio-sitrep.R
\name{check_api_connection_openai}
\alias{check_api_connection_openai}
\title{Check API Connection}
\usage{
check_api_connection_openai(service, api_key)
}
\arguments{
\item{service}{The name of the API service for which the connection is being checked.}

\item{api_key}{The API key used for authentication.}
}
\value{
A logical value indicating whether the connection was successful.
}
\description{
This generic function checks the API connection for a specified service
by dispatching to related methods.
}
