% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{collect_dataframes}
\alias{collect_dataframes}
\title{Collect Dataframes}
\usage{
collect_dataframes()
}
\value{
A character vector of dataframe names.
}
\description{
Collect all the dataframes in the global environment.
}
