% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-azure_openai.R
\name{create_chat_azure_openai}
\alias{create_chat_azure_openai}
\title{Generate text using Azure OpenAI's API}
\usage{
create_chat_azure_openai(
  prompt = list(list(role = "user", content = "Hello")),
  model = Sys.getenv("AZURE_OPENAI_DEPLOYMENT_NAME"),
  api_key = Sys.getenv("AZURE_OPENAI_API_KEY"),
  task = "chat/completions",
  stream = FALSE,
  shiny_session = NULL,
  user_prompt = NULL,
  base_url = Sys.getenv("AZURE_OPENAI_ENDPOINT"),
  api_version = Sys.getenv("AZURE_OPENAI_API_VERSION")
)
}
\arguments{
\item{prompt}{A list of messages to use as the prompt for generating completions.
Each message should be a list with 'role' and 'content' elements.}

\item{model}{A character string for the model to use. Defaults to the Azure OpenAI
deployment name from environment variables if not specified.}

\item{api_key}{A character string for the API key. It will default to the Azure
OpenAI API key from your environment variables if not specified.}

\item{task}{A character string for the API task. Defaults to "chat/completions".}

\item{stream}{Whether to stream the response, defaults to FALSE.}

\item{shiny_session}{A Shiny session object to send messages to the client}

\item{user_prompt}{A user prompt to send to the client}

\item{base_url}{A character string for the base url. It defaults to the Azure
OpenAI endpoint from environment variables if not specified.}

\item{api_version}{A character string for the API version. It defaults to
the Azure OpenAI API version from your environment variables if not specified.}
}
\value{
A list with the generated completions and other information returned
by the API
}
\description{
Use this function to generate text completions using Azure OpenAI's API.
}
