% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-openai.R
\name{create_chat_openai}
\alias{create_chat_openai}
\title{Generate text completions using OpenAI's API for Chat}
\usage{
create_chat_openai(
  prompt = list(list(role = "user", content = "Hello")),
  model = "gpt-4o",
  api_key = Sys.getenv("OPENAI_API_KEY"),
  task = "chat/completions",
  stream = FALSE,
  shiny_session = NULL,
  user_prompt = NULL
)
}
\arguments{
\item{prompt}{The prompt for generating completions}

\item{model}{The model to use for generating text}

\item{api_key}{The API key for accessing OpenAI's API. By default, the
function will try to use the \code{OPENAI_API_KEY} environment variable.}

\item{task}{The task that specifies the API url to use, defaults to
"completions" and "chat/completions" is required for ChatGPT model.}

\item{stream}{Whether to stream the response, defaults to FALSE.}

\item{shiny_session}{A Shiny session object to send messages to the client}

\item{user_prompt}{A user prompt to send to the client}
}
\value{
A list with the generated completions and other information returned
by the API.
}
\description{
Generate text completions using OpenAI's API for Chat
}
\examples{
\dontrun{
openai_create_completion(
  model = "gpt-4o",
  prompt = "Hello world!"
)
}
}
