% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chat.R
\name{create_ide_matching_colors}
\alias{create_ide_matching_colors}
\title{Chat message colors in RStudio}
\usage{
create_ide_matching_colors(role)
}
\arguments{
\item{role}{The role of the message author}
}
\value{
list
}
\description{
This returns a list of color properties for a chat message
}
