% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app.R
\name{create_translator}
\alias{create_translator}
\title{Internationalization for the ChatGPT addin}
\usage{
create_translator(language = getOption("gptstudio.language"))
}
\arguments{
\item{language}{The language to be found in the translation JSON file.}
}
\value{
A Translator from \code{shiny.i18n::Translator}
}
\description{
The language can be set via \code{options("gptstudio.language" = "<language>")}
(defaults to "en") or the "GPTSTUDIO_LANGUAGE" environment variable.
}
