% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-openai_api_calls.R
\name{encode_image}
\alias{encode_image}
\title{Encode an image file to base64}
\usage{
encode_image(image_path)
}
\arguments{
\item{image_path}{String containing the path to the image file}
}
\value{
A base64 encoded string of the image
}
\description{
Encode an image file to base64
}
