% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_all_available_services}
\alias{get_all_available_services}
\title{List supported service providers}
\usage{
get_all_available_services()
}
\value{
A character vector
}
\description{
List the names of all providers/vendors supported by the gptstudio chat app.
}
\examples{
get_all_available_services()


}
