% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_taglist.R
\name{get_node_params}
\alias{get_node_params}
\title{Get Node Parameters}
\usage{
get_node_params(node)
}
\arguments{
\item{node}{A node representing an element or text node in a parsed HTML
document}
}
\value{
A list with the name, attributes, and contents of the node
}
\description{
This function takes a node and returns a list with its name, attributes, and
contents. This functions applies recursively to every element of its contents
until the element is plain text or has no extra content.
}
