% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_taglist.R
\name{get_nodeset_tag_contents}
\alias{get_nodeset_tag_contents}
\title{Get Nodeset Tag Contents}
\usage{
get_nodeset_tag_contents(nodeset)
}
\arguments{
\item{nodeset}{A nodeset representing a parsed HTML document}
}
\value{
A character vector containing the contents of each tag in the nodeset
}
\description{
This function takes a nodeset and returns the contents of each tag.
}
