% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_queries.R
\name{get_selection}
\alias{get_selection}
\title{Wrapper around selectionGet to help with testthat}
\usage{
get_selection()
}
\value{
Text selection via \code{rstudioapi::selectionGet}
}
\description{
Wrapper around selectionGet to help with testthat
}
