% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_queries.R
\name{gpt_chat}
\alias{gpt_chat}
\title{Query an Index}
\usage{
gpt_chat(query, history, style = "tidyverse", skill = "beginner")
}
\arguments{
\item{query}{A character string representing the question or prompt to query
the index with.}

\item{history}{A list of the previous chat responses}

\item{style}{A character string indicating the preferred coding style, the
default is "tidyverse".}

\item{skill}{The self-described skill level of the programmer,
default is "beginner"}
}
\value{
A list containing the instructions for answering the question, the
context in which the question was asked, and the suggested answer.
}
\description{
This function queries an index with a given question or prompt and returns a
set of suggested answers.
}
