% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_queries.R
\name{gpt_chat_in_source}
\alias{gpt_chat_in_source}
\title{ChatGPT in Source}
\usage{
gpt_chat_in_source(
  query,
  history = NULL,
  style = getOption("gptstudio.code_style"),
  skill = getOption("gptstudio.skill")
)
}
\arguments{
\item{query}{A character string representing the question or prompt to query
the index with.}

\item{history}{A list of the previous chat responses}

\item{style}{A character string indicating the preferred coding style, the
default is "tidyverse".}

\item{skill}{The self-described skill level of the programmer,
default is "beginner"}
}
\value{
A list containing the instructions for answering the question, the
context in which the question was asked, and the suggested answer.
}
\description{
Provides the same functionality as \code{gpt_chat()} with minor modifications to
give more useful output in a source (i.e., *.R) file.
}
