% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_queries.R
\name{gpt_create}
\alias{gpt_create}
\title{Use GPT to improve text}
\usage{
gpt_create(
  model,
  temperature,
  max_tokens,
  top_p,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL,
  append_text = TRUE
)
}
\arguments{
\item{model}{The name of the GPT model to use.}

\item{temperature}{A parameter for controlling the randomness of the GPT model's output.}

\item{max_tokens}{Maximum number of tokens to return (related to length of response), defaults to 500}

\item{top_p}{A parameter for controlling the probability of the GPT model's output.}

\item{openai_api_key}{An API key for the OpenAI API.}

\item{openai_organization}{An optional organization ID for the OpenAI API.}

\item{append_text}{Add text to selection rather than replace, defaults to TRUE}
}
\value{
Nothing is returned. The improved text is inserted into the current RStudio session.
}
\description{
This function uses the GPT model from OpenAI to improve the spelling and grammar of the selected text in the current RStudio session.
}
