% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptstudio-package.R
\docType{package}
\name{gptstudio-package}
\alias{gptstudio}
\alias{gptstudio-package}
\title{gptstudio: gptstudio adds functionality based on OpenAI's GPT family of models directly into Rstudio}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

OpenAI offers a number of large language models (LLMs) accessible via API. This package lowers the barrier to using this API inside of RStudio IDE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MichelNivard/gptstudio}
  \item \url{https://michelnivard.github.io/gptstudio/}
  \item Report bugs at \url{https://github.com/MichelNivard/gptstudio/issues}
}

}
\author{
\strong{Maintainer}: James Wade \email{github@jameshwade.com} (\href{https://orcid.org/0000-0002-9740-1905}{ORCID})

Authors:
\itemize{
  \item Michel Nivard \email{m.g.nivard@vu.nl}
}

}
\keyword{internal}
