% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{gptstudio_cache_directory}
\alias{gptstudio_cache_directory}
\title{a function that determines the appropriate directory to cache a token}
\usage{
gptstudio_cache_directory()
}
\description{
a function that determines the appropriate directory to cache a token
}
