% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_chatgpt.R
\name{gptstudio_chat}
\alias{gptstudio_chat}
\title{Run GPTStudio Chat App}
\usage{
gptstudio_chat(host = getOption("shiny.host", "127.0.0.1"))
}
\arguments{
\item{host}{A character string specifying the host on which to run the app.
Defaults to the value of \code{getOption("shiny.host", "127.0.0.1")}.}
}
\value{
This function does not return a value. It runs the Shiny app as a side effect.
}
\description{
Run GPTStudio Chat App
}
\details{
The function performs the following steps:
\enumerate{
\item Verifies that RStudio API is available.
\item Finds an available port for the Shiny app.
\item Creates a temporary directory for the app files.
\item Runs the app as a background job in RStudio.
\item Opens the app in the RStudio viewer pane or browser window.
}
}
\note{
This function is designed to work within the RStudio IDE and requires
the rstudioapi package.
}
\examples{
\dontrun{
gptstudio_chat()
}
}
