% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_chatgpt-in-source.R
\name{gptstudio_chat_in_source_addin}
\alias{gptstudio_chat_in_source_addin}
\title{ChatGPT in Source}
\usage{
gptstudio_chat_in_source_addin()
}
\value{
This function has no return value.
}
\description{
Call this function as a Rstudio addin to ask GPT to improve spelling and
grammar of selected text.
}
\examples{
# Select some text in a source file
# Then call the function as an RStudio addin
\dontrun{
gptstudio_chat_in_source()
}
}
