% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptstudio-sitrep.R
\name{gptstudio_sitrep}
\alias{gptstudio_sitrep}
\title{Current Configuration for gptstudio}
\usage{
gptstudio_sitrep(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value indicating whether to output additional information,
such as API connection checks. Defaults to TRUE.}
}
\value{
Invisibly returns NULL, as the primary purpose of this function is to
print to the console.
}
\description{
This function prints out the current configuration settings for gptstudio and
checks API connections if verbose is TRUE.
}
\examples{
gptstudio_sitrep(verbose = FALSE) # Print basic settings, no API checks
gptstudio_sitrep() # Print settings and check API connections

}
