% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_build_skeleton.R
\name{gptstudio_skeleton_build}
\alias{gptstudio_skeleton_build}
\title{Construct a GPT Studio request skeleton.}
\usage{
gptstudio_skeleton_build(skeleton, skill, style, ...)
}
\arguments{
\item{skeleton}{A GPT Studio request skeleton object.}

\item{skill}{A parameter that affects the style of the chat conversation.}

\item{style}{A parameter that affects the style of the chat conversation.}

\item{...}{Additional arguments.}
}
\value{
An updated GPT Studio request skeleton.
}
\description{
Construct a GPT Studio request skeleton.
}
