% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingface.R
\name{hf_create_completion}
\alias{hf_create_completion}
\title{Generate text completions using HuggingFace's API}
\usage{
hf_create_completion(
  prompt = "",
  model = "gpt2",
  hf_api_key = Sys.getenv("HF_API_KEY")
)
}
\arguments{
\item{prompt}{The prompt for generating completions}

\item{model}{The model to use for generating text}

\item{hf_api_key}{The API key for accessing HuggingFace's API. By default, the
function will try to use the \code{HF_API_KEY} environment variable.}
}
\value{
A list with the generated completions and other information returned
by the API.
}
\description{
Generate text completions using HuggingFace's API
}
\examples{
\dontrun{
hf_create_completion(
  model = "gpt2",
  prompt = "Hello world!"
)
}
}
