% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_taglist.R
\name{html_str_to_nodeset}
\alias{html_str_to_nodeset}
\title{HTML string to xml nodeset}
\usage{
html_str_to_nodeset(str)
}
\arguments{
\item{str}{A character string that represents the HTML to be parsed}
}
\value{
A nodeset representing the parsed HTML
}
\description{
This function takes HTML defined as a string and returns it as a xml_nodeset.
}
