% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_taglist.R
\name{html_to_r}
\alias{html_to_r}
\title{Convert HTML to R Code}
\usage{
html_to_r(html)
}
\arguments{
\item{html}{A character string of HTML code}
}
\value{
A character string of styled R code that can be used to recreate the
HTML structure
}
\description{
This function takes a character string of HTML code and returns a styled R
code that can be used to recreate the HTML structure. The resulting R code is
a character string that can be copied and pasted into an R script or console.
}
