% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_taglist.R
\name{html_to_taglist}
\alias{html_to_taglist}
\title{Convert HTML to a Tag List}
\usage{
html_to_taglist(html)
}
\arguments{
\item{html}{A character string of HTML code}
}
\value{
A tag list that can be used to display the HTML content
}
\description{
This function takes a character string of HTML code and returns a tag list that can be used to
display the HTML content in an R Markdown document or Shiny app. The resulting tag list can be
passed as an argument to the \code{htmltools::tagQuery()} function or used as an input to other HTML
rendering functions in R.
}
