% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_queries.R
\name{insert_text}
\alias{insert_text}
\title{Wrapper around selectionGet to help with testthat}
\usage{
insert_text(improved_text)
}
\arguments{
\item{improved_text}{Text from model queries to inert into script or document}
}
\description{
Wrapper around selectionGet to help with testthat
}
