% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_taglist.R
\name{into_taglist}
\alias{into_taglist}
\title{Paste tags string inside a tagList}
\usage{
into_taglist(tags_str)
}
\arguments{
\item{tags_str}{A list of HTML tags}
}
\value{
A character string that, when evaluated, will produce a tagList object containing the given tags
}
\description{
This function takes a list of HTML tags and returns a character string that, when evaluated,
will produce a tagList object containing the given tags.
}
