% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chat.R
\name{lighten_color}
\alias{lighten_color}
\title{Make a color lighter or darker}
\usage{
lighten_color(color, percentage = 0)
}
\arguments{
\item{color}{An hex color}

\item{percentage}{A number from 0 to 1 indicating how lighter should the color be. When negative it will darken \code{color}}
}
\value{
An hex color
}
\description{
This wraps \code{grDevices::adjustcolor()} for easier usage. Leaves the alpha value as is.
}
