% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app.R
\name{mod_app_ui}
\alias{mod_app_ui}
\title{App UI}
\usage{
mod_app_ui(id, ide_colors = get_ide_theme_info())
}
\arguments{
\item{id}{id of the module}

\item{ide_colors}{List containing the colors of the IDE theme.}
}
\description{
App UI
}
