% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chat.R
\name{mod_chat_ui}
\alias{mod_chat_ui}
\title{Chat UI}
\usage{
mod_chat_ui(
  id,
  translator = create_translator(),
  code_theme_url = get_highlightjs_theme()
)
}
\arguments{
\item{id}{id of the module}

\item{translator}{A Translator from \code{shiny.i18n::Translator}}

\item{code_theme_url}{URL to the highlight.js theme}
}
\description{
Chat UI
}
