% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prompt.R
\name{mod_prompt_server}
\alias{mod_prompt_server}
\title{Prompt Server}
\usage{
mod_prompt_server(id, ide_colors = get_ide_theme_info())
}
\arguments{
\item{id}{id of the module}

\item{ide_colors}{List containing the colors of the IDE theme.}
}
\value{
A shiny server
}
\description{
This server receives the input of the user and makes the chat history
}
