% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_taglist.R
\name{node_params_to_str}
\alias{node_params_to_str}
\title{Convert Node Parameters to String}
\usage{
node_params_to_str(node_params)
}
\arguments{
\item{node_params}{A list of parameters for an HTML tag}
}
\value{
A character string that represents the tag with the given parameters
}
\description{
This function takes a list of parameters for an HTML tag and returns a character string that
represents the tag with the given parameters. Aplies recursively to every child content until
content is text or empty.
}
