% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_gpt.R
\name{openai_chat}
\alias{openai_chat}
\title{Generate an answer to a question using OpenAI}
\usage{
openai_chat(
  model = "text-davinci-003",
  max_tokens = getOption("gptstudio.max_tokens"),
  temperature = 0.5,
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{model}{A string representing the name of the OpenAI model to use for
generating the answer. Default is "text-davinci-003".}

\item{max_tokens}{An integer representing the maximum number of tokens to
generate in the answer. Default is getOption("gptstudio.max_tokens").}

\item{temperature}{A numeric value between 0 and 1 representing the sampling
temperature for the OpenAI model. Default is 0.5.}

\item{openai_api_key}{A string representing the OpenAI API key to use for
making API calls. Default is the value of the environment variable
"OPENAI_API_KEY".}
}
\description{
Generate an answer to a question using OpenAI
}
