% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api_calls.R
\name{openai_create_edit}
\alias{openai_create_edit}
\title{Create edit}
\usage{
openai_create_edit(
  model,
  input = "\\"",
  instruction,
  temperature = 1,
  top_p = 1,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)
}
\arguments{
\item{model}{required; a length one character vector.}

\item{input}{required; defaults to `'"'`; a length one character vector.}

\item{instruction}{required; a length one character vector.}

\item{temperature}{required; defaults to `1`; a length one numeric vector
with the value between `0` and `2`.}

\item{top_p}{required; defaults to `1`; a length one numeric vector with the
value between `0` and `1`.}

\item{openai_api_key}{required; defaults to `Sys.getenv("OPENAI_API_KEY")`
(i.e., the value is retrieved from the `.Renviron` file); a length one
character vector. Specifies OpenAI API key.}

\item{openai_organization}{optional; defaults to `NULL`; a length one
character vector. Specifies OpenAI organization.}
}
\value{
Returns a list, elements of which contain edited version of prompt
  and supplementary information.
}
\description{
#' This code is altered from [openai](https://github.com/irudnyts/openai).
Credit goes to irudnyts for contributions to this function.
}
\details{
Creates an edit based on the provided input, instruction, and parameters. See
[this page](https://beta.openai.com/docs/api-reference/edits/create) for
details.

For arguments description please refer to the official
[docs](https://beta.openai.com/docs/api-reference/edits/create).
}
\examples{
\dontrun{
create_edit(
  model = "text-davinci-edit-001",
  input = "What day of the wek is it?",
  instruction = "Fix the spelling mistakes"
)
}
}
