% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_streaming.R
\name{openai_stream_parse}
\alias{openai_stream_parse}
\title{OpenAI Stream Parse}
\usage{
openai_stream_parse(x)
}
\arguments{
\item{x}{A raw vector representing a chunk of data from the API stream.}
}
\value{
If parsing is successful, a character string of the API response is
returned. In case of an error, an error message is returned instead.
}
\description{
This function handles the streaming data from the OpenAI API.
It concatenates the raw data chunks, attempts to parse JSON and
handles any error messages.
}
\details{
This function was inspired by the \code{{chattr}} R package
(https://github.com/mlverse/chattr).
}
