% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe-audio.R
\name{parse_data_uri}
\alias{parse_data_uri}
\title{Parse a Data URI}
\usage{
parse_data_uri(data_uri)
}
\arguments{
\item{data_uri}{A string. The data URI to parse.}
}
\value{
A list with two elements: 'mime_type' and 'data'.
}
\description{
This function parses a data URI and returns the MIME type and decoded data.
}
