% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{prep_data_prompt}
\alias{prep_data_prompt}
\title{Preps OpenAI model prompt for data documentation}
\usage{
prep_data_prompt(data, method, prompt)
}
\arguments{
\item{data}{A data.frame}

\item{method}{A summarization method, one of "skimr", "glimpse", or "summary"}

\item{prompt}{A prompt string}
}
\value{
A string
}
\description{
Prepares data prompt by summarizing data and printing it
}
\examples{
prep_data_prompt(
  data = mtcars,
  method = "skimr",
  prompt = "This is a test prompt."
)
}
