% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_queries.R
\name{prepare_chat_history}
\alias{prepare_chat_history}
\title{Prepare chat completion prompt}
\usage{
prepare_chat_history(
  history = NULL,
  style = getOption("gptstudio.code_style"),
  skill = getOption("gptstudio.skill")
)
}
\arguments{
\item{history}{A list of previous messages in the conversation (optional).}

\item{style}{The style of the chat conversation (optional). Default is
retrieved from the "gptstudio.code_style" option.}

\item{skill}{The skill to use for the chat conversation (optional). Default
is retrieved from the "gptstudio.skill" option.}
}
\value{
A list containing the body of the request.
}
\description{
This function prepares the chat completion prompt to be sent to the OpenAI API.
}
