% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helpers.R
\name{query_api}
\alias{query_api}
\title{A function that sends an API request and returns the response.}
\usage{
query_api(task, request_body, token = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{task}{A character string that specifies the task to send to the API.}

\item{request_body}{A list that contains the parameters for the task.}

\item{token}{String containing an API key. Defaults to the OPENAI_API_KEY environmental variable if not specified.}
}
\value{
The response from the API.
}
\description{
A function that sends an API request and returns the response.
}
