% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthropic.R
\name{query_api_anthropic}
\alias{query_api_anthropic}
\title{A function that sends a request to the Anthropic API and returns the
response.}
\usage{
query_api_anthropic(request_body, key = Sys.getenv("ANTHROPIC_API_KEY"))
}
\arguments{
\item{request_body}{A list that contains the parameters for the task.}

\item{key}{String containing an Anthropic API key. Defaults
to the ANTHROPIC_API_KEY environmental variable if not specified.}
}
\value{
The response from the API.
}
\description{
A function that sends a request to the Anthropic API and returns the
response.
}
