% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-google.R
\name{query_api_google}
\alias{query_api_google}
\title{A function that sends a request to the Google AI Studio API and returns the
response.}
\usage{
query_api_google(model, request_body, key = Sys.getenv("GOOGLE_API_KEY"))
}
\arguments{
\item{model}{A character string that specifies the model to send to the API.}

\item{request_body}{A list that contains the parameters for the task.}

\item{key}{String containing a Google AI Studio API key. Defaults
to the GOOGLE_API_KEY environmental variable if not specified.}
}
\value{
The response from the API.
}
\description{
A function that sends a request to the Google AI Studio API and returns the
response.
}
