% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingface.R
\name{query_api_huggingface}
\alias{query_api_huggingface}
\title{A function that sends a request to the HuggingFace API and returns the
response.}
\usage{
query_api_huggingface(task, request_body, token = Sys.getenv("HF_API_KEY"))
}
\arguments{
\item{task}{A character string that specifies the task to send to the API.}

\item{request_body}{A list that contains the parameters for the task.}

\item{token}{String containing a HuggingFace API key. Defaults
to the HF_API_KEY environmental variable if not specified.}
}
\value{
The response from the API.
}
\description{
A function that sends a request to the HuggingFace API and returns the
response.
}
