% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-perplexity.R
\name{query_api_perplexity}
\alias{query_api_perplexity}
\title{Send a request to the Perplexity API and return the response}
\usage{
query_api_perplexity(request_body, api_key = Sys.getenv("PERPLEXITY_API_KEY"))
}
\arguments{
\item{request_body}{A list containing the body of the POST request.}

\item{api_key}{String containing a Perplexity API key. Defaults to the
PERPLEXITY_API_KEY environmental variable if not specified.}
}
\value{
A parsed JSON object as the API response.
}
\description{
This function sends a JSON post request to the Perplexity API,
retries on failure up to three times, and returns the response.
The function handles errors by providing a descriptive message and failing gracefully.
}
