% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api_calls.R
\name{query_openai_api}
\alias{query_openai_api}
\title{A function that sends a request to the OpenAI API and returns the response.}
\usage{
query_openai_api(
  task,
  request_body,
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{task}{A character string that specifies the task to send to the API.}

\item{request_body}{A list that contains the parameters for the task.}

\item{openai_api_key}{String containing an OpenAI API key. Defaults to the OPENAI_API_KEY environmental variable if not specified.}
}
\value{
The response from the API.
}
\description{
A function that sends a request to the OpenAI API and returns the response.
}
