% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_chat_gpt.R
\name{random_port}
\alias{random_port}
\title{Generate a random safe port number}
\usage{
random_port()
}
\value{
A single integer representing the randomly selected safe port number.
}
\description{
This function generates a random port number between 3000 and 8000, but excludes some commonly unsafe ports from the selection.
}
\examples{
random_port()
}
