% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helpers.R, R/openai_api_calls.R
\name{request_base}
\alias{request_base}
\title{Base for a request to an API service}
\usage{
request_base(task, token = Sys.getenv("OPENAI_API_KEY"))

request_base(task, token = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{task}{character string specifying an OpenAI API endpoint task}

\item{token}{String containing an OpenAI API key. Defaults to the OPENAI_API_KEY environmental variable if not specified.}
}
\value{
An httr2 request object

An httr2 request object
}
\description{
This function sends a request to a specific API \code{task} endpoint at the base URL and authenticates with an API key using a Bearer token. The default is OpenAI's API using a base URL of \code{https://api.openai.com/v1}.

This function sends a request to a specific OpenAI API \code{task} endpoint at the base URL \code{https://api.openai.com/v1}, and authenticates with an API key using a Bearer token.
}
\keyword{api,}
\keyword{authentication}
\keyword{openai,}
