% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthropic.R
\name{request_base_anthropic}
\alias{request_base_anthropic}
\title{Base for a request to the Anthropic API}
\usage{
request_base_anthropic(key = Sys.getenv("ANTHROPIC_API_KEY"))
}
\arguments{
\item{key}{String containing an Anthropic API key. Defaults to the
ANTHROPIC_API_KEY environmental variable if not specified.}
}
\value{
An httr2 request object
}
\description{
This function sends a request to the Anthropic API endpoint and
authenticates with an API key.
}
