% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-cohere.R
\name{request_base_cohere}
\alias{request_base_cohere}
\title{Base for a request to the Cohere Chat API}
\usage{
request_base_cohere(api_key = Sys.getenv("COHERE_API_KEY"))
}
\arguments{
\item{api_key}{String containing a Cohere API key. Defaults to the
COHERE_API_KEY environment variable if not specified.}
}
\value{
An \code{httr2} request object pre-configured with the API endpoint and required headers.
}
\description{
This function sets up a POST request to the Cohere Chat API's chat endpoint
and includes necessary headers such as 'accept', 'content-type', and 'Authorization'
with a bearer token.
}
