% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-google.R
\name{request_base_google}
\alias{request_base_google}
\title{Base for a request to the Google AI Studio API}
\usage{
request_base_google(model, key = Sys.getenv("GOOGLE_API_KEY"))
}
\arguments{
\item{model}{character string specifying a Google AI Studio API model}

\item{key}{String containing a Google AI Studio API key. Defaults to the
GOOGLE_API_KEY environmental variable if not specified.}
}
\value{
An httr2 request object
}
\description{
This function sends a request to a specific Google AI Studio API endpoint and
authenticates with an API key.
}
