% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingface.R
\name{request_base_huggingface}
\alias{request_base_huggingface}
\title{Base for a request to the HuggingFace API}
\usage{
request_base_huggingface(task, token = Sys.getenv("HF_API_KEY"))
}
\arguments{
\item{task}{character string specifying a HuggingFace API endpoint task}

\item{token}{String containing a HuggingFace API key. Defaults to the
HF_API_KEY environmental variable if not specified.}
}
\value{
An httr2 request object
}
\description{
This function sends a request to a specific HuggingFace API endpoint and
authenticates with an API key using a Bearer token.
}
