% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_palm_makersuite.R
\name{request_base_palm}
\alias{request_base_palm}
\title{Base for a request to the PALM (MakerSuite) API}
\usage{
request_base_palm(model, key = Sys.getenv("PALM_API_KEY"))
}
\arguments{
\item{model}{character string specifying a PALM (MakerSuite) API model}

\item{key}{String containing a PALM (MakerSuite) API key. Defaults to the
PALM_API_KEY environmental variable if not specified.}
}
\value{
An httr2 request object
}
\description{
This function sends a request to a specific PALM (MakerSuite) API endpoint and
authenticates with an API key.
}
