% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-perplexity.R
\name{request_base_perplexity}
\alias{request_base_perplexity}
\title{Base for a request to the Perplexity API}
\usage{
request_base_perplexity(api_key = Sys.getenv("PERPLEXITY_API_KEY"))
}
\arguments{
\item{api_key}{String containing a Perplexity API key. Defaults to the
PERPLEXITY_API_KEY environment variable if not specified.}
}
\value{
An \code{httr2} request object pre-configured with the API endpoint and required headers.
}
\description{
This function sets up a POST request to the Perplexity API's chat/completions endpoint
and includes necessary headers such as 'accept', 'content-type', and 'Authorization'
with a bearer token.
}
