% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app.R
\name{rgb_str_to_hex}
\alias{rgb_str_to_hex}
\title{RGB str to hex}
\usage{
rgb_str_to_hex(rgb_string)
}
\arguments{
\item{rgb_string}{The RGB string as returned by \code{rstudioapi::getThemeInfo()}}
}
\value{
hex color
}
\description{
RGB str to hex
}
