% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{roxygenAddin}
\alias{roxygenAddin}
\title{Add Roxygen documentation to a function}
\usage{
roxygenAddin()
}
\value{
NULL (nothing is returned; the generated roxygen skeleton is inserted into the document).
}
\description{
This function uses the OpenAI API to generate a roxygen skeleton for the
current selection in RStudio. The roxygen skeleton is then inserted into the
 document.
}
