% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chatgpt_app.R
\name{run_chatgpt_app}
\alias{run_chatgpt_app}
\title{Run the ChatGPT app}
\usage{
run_chatgpt_app()
}
\value{
Nothing.
}
\description{
This starts the chatgpt app. It is exported to be able to run it from an R script.
}
