% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_freeform_editor.R
\name{run_gpt_freeform}
\alias{run_gpt_freeform}
\title{Shiny app that supports GPT freeform addin}
\usage{
run_gpt_freeform()
}
\value{
Nothing is returned, a shiny app is run
}
\description{
Shiny app that supports GPT freeform addin
}
