% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandgAddin.R
\name{sandgAddin}
\alias{sandgAddin}
\title{spelling and grammar Addin}
\usage{
sandgAddin()
}
\description{
Call this function as a Rstudio addin to ask GPT to improve spelling and
grammar of selected text.
}
