% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{set_allowed_models}
\alias{set_allowed_models}
\title{Set allowed models by provider}
\usage{
set_allowed_models(service, models = NULL)
}
\arguments{
\item{service}{The API service}

\item{models}{A character vector containing the list of allowed models that should be shown
in the dropdown selector. If \code{NULL} (default), all models will be available.}
}
\description{
Sometimes it is useful to restrict the list of options in the model selection dropdown of the
chat app. This function will check against \code{get_available_models()} to restrict the list to
models that are actually available.
}
