% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{set_openai_api_key}
\alias{set_openai_api_key}
\title{Set API Key to Environment Variable}
\usage{
set_openai_api_key()
}
\description{
This function sets the `OPENAI_API_KEY` environment variable to the
OPENAI_API_KEY if it is valid.
}
