% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_gpt.R
\name{setup_gpt_chat}
\alias{setup_gpt_chat}
\title{Setup File for GPT Chat}
\usage{
setup_gpt_chat(path = "gpt_q_and_a.qmd")
}
\arguments{
\item{path}{The path of the file to be created. Defaults to "gpt_q_and_a.qmd"}
}
\description{
Setup File for GPT Chat
}
