% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api.R
\name{simple_api_check}
\alias{simple_api_check}
\title{Send a simple API request}
\usage{
simple_api_check(api_key)
}
\arguments{
\item{api_key}{An API key for OpenAI's API.}
}
\value{
The status code of the API request.
}
\description{
Send a simple API request
}
