% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamHandler.R
\name{stream_chat_completion}
\alias{stream_chat_completion}
\title{Stream Chat Completion}
\usage{
stream_chat_completion(
  prompt,
  element_callback = cat,
  model = getOption("gptstudio.chat_model"),
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{prompt}{The prepared chat completion prompt.}

\item{element_callback}{A callback function to handle each element of the
streamed response (optional).}

\item{model}{The model to use for chat completion (optional). Default is
"gpt-3.5-turbo".}

\item{openai_api_key}{The OpenAI API key (optional). By default, it is
fetched from the "OPENAI_API_KEY" environment variable.}
}
\value{
the same as \code{curl::curl_fetch_stream}
}
\description{
This function sends the prepared chat completion request to the OpenAI API and
retrieves the streamed response.
}
