% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingMessage.R
\name{streamingMessage}
\alias{streamingMessage}
\title{Streaming message}
\usage{
streamingMessage(
  ide_colors = get_ide_theme_info(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Places an invisible empty chat message that will hold a streaming message.
It can be resetted dynamically inside a shiny app
}
