% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prompt.R
\name{style_chat_message}
\alias{style_chat_message}
\title{Style chat message}
\usage{
style_chat_message(message)
}
\arguments{
\item{message}{A chat message.}
}
\value{
An HTML element.
}
\description{
Style a message based on the role of its author.
}
