% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize data}
\usage{
summarize_data(
  data,
  method = c("skimr", "skimr_lite", "column_types", "summary")
)
}
\arguments{
\item{data}{A data frame}

\item{method}{A character vector specifying the method to use for
summarizing the data. Must be one of "skimr", "skimr_lite", "column_types",
or "summary". Default is "skimr".}
}
\value{
Summarized data according to specified method
}
\description{
Summarize a data frame using one of three methods.
}
