% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_queries.R
\name{warn_about_openai_pkg}
\alias{warn_about_openai_pkg}
\title{Warn about openai package}
\usage{
warn_about_openai_pkg()
}
\description{
Warn the user if the openai package is not installed and their R version is
compatible with the current release of the package.
}
\examples{
\dontrun{
warn_about_openai_pkg()
}
}
