% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcomeMessage.R
\name{welcomeMessage}
\alias{welcomeMessage}
\title{Welcome message}
\usage{
welcomeMessage(
  ide_colors = get_ide_theme_info(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
HTML widget for showing a welcome message in the chat app.
This has been created to be able to bind the message to a shiny event to trigger a new render.
}
