% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prompt.R
\name{chat_create_system_prompt}
\alias{chat_create_system_prompt}
\title{Create system prompt}
\usage{
chat_create_system_prompt(
  style = c("tidyverse", "base", "no preference", NULL),
  skill = c("beginner", "intermediate", "advanced", "genius", NULL),
  task = c("coding", "general", "advanced developer", "custom"),
  custom_prompt = NULL,
  in_source
)
}
\arguments{
\item{style}{A character string indicating the preferred coding style, the
default is "tidyverse".}

\item{skill}{The self-described skill level of the programmer,
default is "beginner"}

\item{task}{The task to be performed: "coding", "general", or "advanced developer".}

\item{custom_prompt}{An optional custom prompt to be displayed.}

\item{in_source}{Whether to add instructions to act as in a source script.}
}
\value{
A string
}
\description{
This creates a system prompt based on the user defined parameters.
}
