% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_chat_style.R
\name{chat_history_append}
\alias{chat_history_append}
\title{Append to chat history}
\usage{
chat_history_append(history, role, content)
}
\arguments{
\item{history}{List containing previous responses.}

\item{role}{Author of the message. One of \code{c("user", "assistant")}}

\item{content}{Content of the message. If it is from the user most probably
comes from an interactive input.}
}
\value{
list of chat messages
}
\description{
This appends a new response to the chat history
}
