% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcomeMessage.R
\name{chat_message_default}
\alias{chat_message_default}
\title{Default chat message}
\usage{
chat_message_default(translator = create_translator())
}
\arguments{
\item{translator}{A Translator from \code{shiny.i18n::Translator}}
}
\value{
A default chat message for welcoming users.
}
\description{
Default chat message
}
