% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app.R
\name{create_chat_app_theme}
\alias{create_chat_app_theme}
\title{Chat App Theme}
\usage{
create_chat_app_theme(ide_colors = get_ide_theme_info())
}
\arguments{
\item{ide_colors}{List containing the colors of the IDE theme.}
}
\value{
A bslib theme
}
\description{
Create a bslib theme that matches the user's RStudio IDE theme.
}
