% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_chat_style.R
\name{create_ide_matching_colors}
\alias{create_ide_matching_colors}
\title{Chat message colors in RStudio}
\usage{
create_ide_matching_colors(role, ide_colors = get_ide_theme_info())
}
\arguments{
\item{role}{The role of the message author}

\item{ide_colors}{List containing the colors of the IDE theme.}
}
\value{
list
}
\description{
This returns a list of color properties for a chat message
}
