% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-openai.R
\name{get_available_endpoints}
\alias{get_available_endpoints}
\title{List supported endpoints}
\usage{
get_available_endpoints()
}
\value{
A character vector
}
\description{
Get a list of the endpoints supported by gptstudio.
}
\examples{
get_available_endpoints()
}
