% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api_calls.R
\name{get_available_models}
\alias{get_available_models}
\title{List supported models}
\usage{
get_available_models(service)
}
\arguments{
\item{service}{The API service}
}
\value{
A character vector
}
\description{
Get a list of the models supported by the OpenAI API.
}
\examples{
get_available_endpoints()
}
