% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app.R
\name{get_ide_theme_info}
\alias{get_ide_theme_info}
\title{Get IDE theme information.}
\usage{
get_ide_theme_info()
}
\value{
A list with three components:
\item{is_dark}{A boolean indicating whether the current IDE theme is dark.}
\item{bg}{The current IDE theme's background color.}
\item{fg}{The current IDE theme's foreground color.}
}
\description{
This function returns a list with the current IDE theme's information.
}
