% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptstudio-package.R
\docType{package}
\name{gptstudio-package}
\alias{gptstudio}
\alias{gptstudio-package}
\title{gptstudio: Use Large Language Models Directly in your Development Environment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Large language models are readily accessible via API. This package lowers the barrier to use the API inside of your development environment. For more on the API, see \url{https://platform.openai.com/docs/introduction}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MichelNivard/gptstudio}
  \item \url{https://michelnivard.github.io/gptstudio/}
  \item Report bugs at \url{https://github.com/MichelNivard/gptstudio/issues}
}

}
\author{
\strong{Maintainer}: James Wade \email{github@jameshwade.com} (\href{https://orcid.org/0000-0002-9740-1905}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Michel Nivard \email{m.g.nivard@vu.nl} [copyright holder]
}

}
\keyword{internal}
