% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_build_skeleton.R
\name{gptstudio_skeleton_build}
\alias{gptstudio_skeleton_build}
\title{Construct a GPT Studio request skeleton.}
\usage{
gptstudio_skeleton_build(skeleton, skill, style, task, custom_prompt, ...)
}
\arguments{
\item{skeleton}{A GPT Studio request skeleton object.}

\item{skill}{The skill level of the user for the chat conversation. This can
be set through the "gptstudio.skill" option. Default is the
"gptstudio.skill" option. Options are "beginner", "intermediate",
"advanced", and "genius".}

\item{style}{The style of code to use. Applicable styles can be retrieved
from the "gptstudio.code_style" option. Default is the
"gptstudio.code_style" option. Options are "base", "tidyverse", or "no
preference".}

\item{task}{Specifies the task that the assistant will help with. Default is
"coding". Others are "general", "advanced developer", and "custom".}

\item{custom_prompt}{This is a custom prompt that may be used to guide the AI
in its responses. Default is NULL. It will be the only content provided to
the system prompt.}

\item{...}{Additional arguments.}
}
\value{
An updated GPT Studio request skeleton.
}
\description{
Construct a GPT Studio request skeleton.
}
