% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chat.R
\name{mod_chat_server}
\alias{mod_chat_server}
\title{Chat server}
\usage{
mod_chat_server(
  id,
  ide_colors = get_ide_theme_info(),
  translator = create_translator(),
  settings,
  history
)
}
\arguments{
\item{id}{id of the module}

\item{ide_colors}{List containing the colors of the IDE theme.}

\item{translator}{Translator from \code{shiny.i18n::Translator}}

\item{settings, history}{Reactive values from the settings and history module}
}
\description{
Chat server
}
