% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prompt.R
\name{mod_prompt_server}
\alias{mod_prompt_server}
\title{Prompt Server}
\usage{
mod_prompt_server(id)
}
\arguments{
\item{id}{id of the module}
}
\value{
A shiny server
}
\description{
This server receives the input of the user and makes the chat history
}
