% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prompt.R
\name{mod_prompt_ui}
\alias{mod_prompt_ui}
\title{Prompt of the chat}
\usage{
mod_prompt_ui(id, translator = create_translator())
}
\arguments{
\item{id}{id of the module}

\item{translator}{A Translator from \code{shiny.i18n::Translator}}
}
\value{
HTML element
}
\description{
HTML element with a text input and some buttons
}
