% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_chatgpt.R
\name{random_port}
\alias{random_port}
\title{Generate a random safe port number}
\usage{
random_port()
}
\value{
A single integer representing the randomly selected safe port number.
}
\description{
This function generates a random port allowed by shiny::runApp.
}
