% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chatgpt_app.R
\name{run_chatgpt_app}
\alias{run_chatgpt_app}
\title{Run the ChatGPT app}
\usage{
run_chatgpt_app(
  ide_colors = get_ide_theme_info(),
  host = getOption("shiny.host", "127.0.0.1"),
  port = getOption("shiny.port")
)
}
\arguments{
\item{ide_colors}{List containing the colors of the IDE theme.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port between 3000:8000, excluding ports that are blocked
by Google Chrome for being considered unsafe: 3659, 4045, 5060,
5061, 6000, 6566, 6665:6669 and 6697. Up to twenty random
ports will be tried.}
}
\value{
Nothing.
}
\description{
This starts the chatgpt app. It is exported to be able to run it from an R
script.
}
