% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingMessage.R
\name{streamingMessage}
\alias{streamingMessage}
\title{Streaming message}
\usage{
streamingMessage(
  ide_colors = get_ide_theme_info(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{ide_colors}{List containing the colors of the IDE theme.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{The element's id}
}
\description{
Places an invisible empty chat message that will hold a streaming message.
It can be reset dynamically inside a shiny app
}
