% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_chat_style.R
\name{style_chat_message}
\alias{style_chat_message}
\title{Style chat message}
\usage{
style_chat_message(message, ide_colors = get_ide_theme_info())
}
\arguments{
\item{message}{A chat message.}

\item{ide_colors}{List containing the colors of the IDE theme.}
}
\value{
An HTML element.
}
\description{
Style a message based on the role of its author.
}
