% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcomeMessage.R
\name{welcomeMessage}
\alias{welcomeMessage}
\title{Welcome message}
\usage{
welcomeMessage(
  ide_colors = get_ide_theme_info(),
  translator = create_translator(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{ide_colors}{List containing the colors of the IDE theme.}

\item{translator}{A Translator from \code{shiny.i18n::Translator}}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{The element's id}
}
\description{
HTML widget for showing a welcome message in the chat app.
This has been created to be able to bind the message to a shiny event to trigger a new render.
}
